split the weight chart file into components
